requirements.txt
opencv-contrib-python
numpy
pillow
pillow-heif
imutils
scipy